+++ src/lua/creature/creatureevent.cpp
@@ -430,9 +430,9 @@ void CreatureEvent::executeHealthChange(Creature* creature, Creature* attacker, 
 	if (getScriptInterface()->protectedCall(L, 7, 4) != 0) {
 		LuaScriptInterface::reportError(nullptr, LuaScriptInterface::popString(L));
 	} else {
-		damage.primary.value = std::abs(LuaScriptInterface::getNumber<int32_t>(L, -4));
+		damage.primary.value = std::abs(LuaScriptInterface::getNumber<int64_t>(L, -4));
 		damage.primary.type = LuaScriptInterface::getNumber<CombatType_t>(L, -3);
-		damage.secondary.value = std::abs(LuaScriptInterface::getNumber<int32_t>(L, -2));
+		damage.secondary.value = std::abs(LuaScriptInterface::getNumber<int64_t>(L, -2));
 		damage.secondary.type = LuaScriptInterface::getNumber<CombatType_t>(L, -1);
 
 		lua_pop(L, 4);
@@ -475,9 +475,9 @@ void CreatureEvent::executeManaChange(Creature* creature, Creature* attacker, Co
 	if (getScriptInterface()->protectedCall(L, 7, 4) != 0) {
 		LuaScriptInterface::reportError(nullptr, LuaScriptInterface::popString(L));
 	} else {
-		damage.primary.value = LuaScriptInterface::getNumber<int32_t>(L, -4);
+		damage.primary.value = LuaScriptInterface::getNumber<int64_t>(L, -4);
 		damage.primary.type = LuaScriptInterface::getNumber<CombatType_t>(L, -3);
-		damage.secondary.value = LuaScriptInterface::getNumber<int32_t>(L, -2);
+		damage.secondary.value = LuaScriptInterface::getNumber<int64_t>(L, -2);
 		damage.secondary.type = LuaScriptInterface::getNumber<CombatType_t>(L, -1);
 		lua_pop(L, 4);
 	}
