+++ src/lua/functions/core/game/global_functions.cpp
@@ -297,7 +297,7 @@ int GlobalFunctions::luaDoAreaCombatHealth(lua_State* L) {
 		CombatDamage damage;
 		damage.origin = getNumber<CombatOrigin>(L, 8, ORIGIN_SPELL);
 		damage.primary.type = combatType;
-		damage.primary.value = normal_random(getNumber<int32_t>(L, 6), getNumber<int32_t>(L, 5));
+		damage.primary.value = normal_random(getNumber<int64_t>(L, 6), getNumber<int64_t>(L, 5));
 
 		Combat::doCombatHealth(creature, getPosition(L, 3), area, damage, params);
 		pushBoolean(L, true);
@@ -333,7 +333,7 @@ int GlobalFunctions::luaDoTargetCombatHealth(lua_State* L) {
 	CombatDamage damage;
 	damage.origin = getNumber<CombatOrigin>(L, 7, ORIGIN_SPELL);
 	damage.primary.type = combatType;
-	damage.primary.value = normal_random(getNumber<int32_t>(L, 4), getNumber<int32_t>(L, 5));
+	damage.primary.value = normal_random(getNumber<int64_t>(L, 4), getNumber<int64_t>(L, 5));
 
 	// Check if it's a healing then we sould add the non-aggresive tag
 	if (combatType == COMBAT_HEALING || (combatType == COMBAT_MANADRAIN && damage.primary.value > 0)) {
@@ -363,7 +363,7 @@ int GlobalFunctions::luaDoAreaCombatMana(lua_State* L) {
 		CombatDamage damage;
 		damage.origin = getNumber<CombatOrigin>(L, 7, ORIGIN_SPELL);
 		damage.primary.type = COMBAT_MANADRAIN;
-		damage.primary.value = normal_random(getNumber<int32_t>(L, 4), getNumber<int32_t>(L, 5));
+		damage.primary.value = normal_random(getNumber<int64_t>(L, 4), getNumber<int64_t>(L, 5));
 
 		Position pos = getPosition(L, 2);
 		Combat::doCombatMana(creature, pos, area, damage, params);
@@ -397,7 +397,7 @@ int GlobalFunctions::luaDoTargetCombatMana(lua_State* L) {
 	CombatDamage damage;
 	damage.origin = getNumber<CombatOrigin>(L, 6, ORIGIN_SPELL);
 	damage.primary.type = COMBAT_MANADRAIN;
-	damage.primary.value = normal_random(getNumber<int32_t>(L, 3), getNumber<int32_t>(L, 4));
+	damage.primary.value = normal_random(getNumber<int64_t>(L, 3), getNumber<int64_t>(L, 4));
 
 	Combat::doCombatMana(creature, target, damage, params);
 	pushBoolean(L, true);
